+++ CHANGELOG.md
@@ -6,6 +6,21 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 ### Added
+- Docstrings and doctestable examples to `record.py`.
+- Inputs can be validated using operations
+  - `validate` parameter in `Input` takes `Operation.instance_name`
+- Logistic Regression with SAG optimizer
+- Test tensorflow DNNEstimator documentation exaples in CI
+- Add python code for tensorflow DNNEstimator
+- Ability to run a subflow as if it were an operation using the
+  `dffml.dataflow.run` operation.
+### Fixed
+- New model tutorial mentions file paths that should be edited.
+- DataFlow is no longer a dataclass to prevent it from being exported
+  incorrectly.
+
+## [0.3.5] - 2020-03-10
+### Added
 - Parent flows can now forward inputs to active contexts of subflows.
   - `forward` parameter in `DataFlow`
   - `subflow` in `OperationImplementationContext`
@@ -30,6 +45,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - config directories and files associated with ConfigLoaders have been renamed
   to configloader.
 - Model config directory parameters are now `pathlib.Path` objects
+- New model tutorial and `skel/model` use simplifeid model API.
 
 ## [0.3.4] - 2020-02-28
 ### Added